requirements.txt
google-genai
python-dotenv==1.0.0